Moved WP_ADMIN_USER, WP_USERNAME and WP_PASSWORD into config file.
