+++ test/e2e/support/utils/login.js
@@ -3,16 +3,7 @@
  */
 import { pressWithModifier } from './press-with-modifier';
 import { META_KEY } from './';
-
-const WP_ADMIN_USER = {
-	username: 'admin',
-	password: 'password',
-};
-
-const {
-	WP_USERNAME = WP_ADMIN_USER.username,
-	WP_PASSWORD = WP_ADMIN_USER.password,
-} = process.env;
+import { WP_USERNAME, WP_PASSWORD } from './config';
 
 export async function login( username = WP_USERNAME, password = WP_PASSWORD ) {
 	await page.focus( '#user_login' );
