+++ test/e2e/support/utils/switch-to-admin-user.js
@@ -3,15 +3,7 @@
  */
 import { goToWPPath } from './go-to-wp-path';
 import { login } from './login';
-
-const WP_ADMIN_USER = {
-	username: 'admin',
-	password: 'password',
-};
-
-const {
-	WP_USERNAME = WP_ADMIN_USER.username,
-} = process.env;
+import { WP_USERNAME, WP_ADMIN_USER } from './config';
 
 /**
  * Switches the current user to the admin user (if the user
